Add files via upload
This was a map that a colleague and I used for the Big Data Hackathon 2017. Our main goal for our application was to create a fully dynamic mapping application that showed the public trends in mental development disorders. These disorders ranged from personality disorders to intellectual disorders. A researcher and author Richard Louv has a theory he created that he calls Nature Deficit Disorder. We wanted to test this theory out and see if Louv is crazy or not, and while comparing open - space/ greenery to mental development disorders we can predict general trends throughout the subregions of San Diego County. 

Our results actually proved Louv's theory to be true. The subregions (South Eastern San Diego, Harbison Crest/ El Cajon, and Chula Vista)  that had the least amount of green space located within their boundaries had the highest cases of mental development, and the subregions that had the highest amount of green space had the lowest cases of mental development. Now this also ha